Update Godot to support clang 6 compilation
Clang 6 is the default clang compiler on Ubuntu 18.04 Bionic.
The previous Godot commit used was not compiling due to a bug in
Clang 6 that has been corrected in Clang 7 development branch.

A workaround has been integrated in Godot [1] in the current
development branch and was backported in the branch 2.1. It
has not yet been integrated in the branch 3.x.

Several include files in Godot are now generated by the build system.
To avoid having to recreate the rules to generate these files with
Bazel, we picked a commit that contained the workaround for Clang 6
but that does not integrate these automatically generated include
files.

[1] https://github.com/godotengine/godot/pull/18868