Install Python extensions with the same name to support multiple Python versions
We relax the conditions under which an error message is thrown if multiple
libraries with the same name are installed. The error message is only thrown
if the library file name starts with the prefix `lib`. Python C/C++ extensions
file names do not start with this prefix. This allows to install these files
in different subfolders without any problem.

The limitation is that we assume that no library depends on these Python C/C++
extensions and therefore nothing requires any RPath fix-up. This limitation
is due to the fact that these potentially duplicated files are not added to
the dictionary which maps the file name to its install location which is used
in the RPath fixup step.