automotive: Remove car_sim_lcm (#9474)
This demo uses RigidBodyPlant, which is heading for deprecation.  We can
resurrect this demo using MultibodyPlant in the future, if we wish.  The
current code isn't very useful either as a standalone example, or a
reference for doing the Multibody porting, so it's not worth moving into
the attic, either.