Add basis of geometry roles
Roles are defined by role-related properties. This includes the underlying
class that supports geometry properties (and tests) and the declaration of
specific roles (currently only proximity and illustration).

The roles are not incorporated anywhere and do not affect functionality or
API. It is merely the basis for defining roles which will *eventually*
enter the API.