framework: Improve error for missing abstract input model_value
The model_value was documented to be (almost always) required, but in
practice many systems were omitting it.  In preparation for enforcing
that its required, we need to make the diagnostic message when its
missing much more useful, so that users have some hope of isolating the
new errors that will appear.