manipulation/dev: Remove misleading install_data rule
There was an install_data rule declared but it didn't install anything
because it wasn't mentioned in the root-level install manifest.  The
install linter didn't catch this problem because dev packages were
excluded from the linting, because they're not supposed to be installed.

Even though the models were declared with public visibility previously,
here we demote them to private, since they do not seem to be curated in
a way where they are intended for external use.