libdrake: Document indirectly-installed components
Some targets are nominally excluded from the install (because they are
private visibility, or part of examples, or etc.) but are dependencies
of installed public code.  We should express them as comments in the
installed components list, so that authors and reviewers can be sure
that they understand what's being installed.