Use ExpressionAddFactory to Expand instead of accumulate
So far, we have used `std::accumulate` and `operator+` for summations in
`.Expand()`. This is not efficient as `operator+` creates a new Expression every
time it's used. Instead, we use `ExpressionAddFactory` to collect summands and
use `.ToExpression()` at the end to generate a sum.

It gives a performance improvement. For example, `pow(x + y + z, 50).Expand()`
took 43.784 sec on an iMac27 (Late 2015, 4Ghz Intel Core i7, 32 GB Ram). Using
this commit, it takes 1.342 sec on the same machine.