systems/controllers: Fork inverse_dynamics into attic
The inverse dynamics code has disjunct implementations supporting both
RigidBodyTree and MultibodyPlant.  Instead of having one class with
if-else inside every method, we should have two classes.  A future
commit will rework the implementations; this commit merely copies the
files, so that the subsequent edits are clear in the history.