Add Context-modifying API
Extend SceneGraph API to support operations that modify the geometry
state stored in a context instead of in the SceneGraph instance's model.
These changes are currently limited to geometry and geometry-related
functions. (Changes to frames and sources are deferred because they will
affect the SceneGraph's ports and we want to make sure we have some
coordination mechanism appropriately planned).

Extended SceneGraph documentation to call out the two API variants.

APIs include:
  1. RegisterGeometry (onto frame and onto other geometry)
  2. RemoveGeometry
  3. Add collision filters (ExcludeCollisionsWithin and
     ExcludeCollisionsBetween)
  4. SceneGraphInspector extended to provide more introspection
     functionality. Supporting functionality also pushed into
     GeometryState and ProximityEngine.