geometry: SignedDistanceToPoint supports Sphere and Box.
Fix a memory bug that showed up with clang but not gcc.

Restructure unit tests.

Restructure unit tests to use value-parameterized tests.

Change the functor DistanceToPoint from "struct" to "class."

Set tolerance when a query point is near the center of a sphere or near
the boundary of a box. Test rigid transforms of a sphere or a box.
Improve readability of the code to classify a query point as outside or
inside or on the boundary of a box.

Update documentation.

Improved readability of clamp(). Added more threshold tests.