tools: Don't force -Werror when compiling Drake as an external
We can't really know what compiler is being used, and so we might fail
due to spurious warnings.  This is OK if Drake was the main workspace,
because the user can opt-out, but with drake as a Bazel external, there
is no easy way for a user to set Drake-specific CXXFLAGS to opt-out.

This also more closely matches best practice to only force -Werror
during development, but not for source releases.