multibody/tree: Rework build files from move
We are simplifying the directory structure.

This commit reworks the build system and pathnames to reflect the new
directory location.  Almost no out-of-package using-code is ported to
the new paths yet -- other than a single benchmark test case where it
was easier to update it than to add test library forwarders.