Add RemoveRole API to dev SceneGraph
Extends the dev/SceneGraph to give the ability to remove illustration or
perception roles from particular geometries (or to all of the
relevant geometries affixed to a frame).

Because this is dev/SceneGraph, it does _not_ enable the ability
to remove proximity roles (as dev/SceneGraph doesn't actually do
proximity).

This should be considered a stop-gap while the full code comes in to
play.