Add new logic to texture generation
For meshes very particularly, if no `diffuse_map` property is defined,
RenderEngineVtk will attempt to look for a png file of the same name
and in the same folder as a texture candidate.

Note: this revealed a bug in the engine cloning -- textures were not
cloned properly; that has now been repaired.

This should be considered a short-term hack.