framework: SystemConstraint knows its System (#10407)
This brings it in line with other indexed System helpers such as
InputPort, OutputPort, etc.  It will also enable the constraint
to pass the `const System&` being evaluated to a functor, so that
the functor doesn't have to capture the System pointer and thus
can be instance-agnostic.

Similarly, SystemConstraint is also now marked `final`.  Since it did
not have a virtual destructor anyway, this was always its intent; it's
just an omission from its inception that the keyword was omitted.
Being final is also consistent with other helpers (InputPort, etc.).