systems/lcm: Enable discrete updates with abstract output
If we know the message size ahead of time, we can still store it in
discrete state, even if the output port is message-valued.

Port kuka_iiwa_arm and schunk_wsg use of LcmSubscriberSystem to eschew
LcmAndVectorBaseTranslator in favor of this new approach, and deprecate
all of the LcmAndVectorBaseTranslator glue for those two components.

This reverts portions of 145b873d7098aa10338b9c700941a9ed7c2c52ae.