Fixes error in geometry removal (with proximity role)
GeometryState maintains a map between ProximityIndex and GeometryIndex
for dynamic geometries. When a dynamic geometry with a proximity role
is removed, that map should decrease in size.

Prior to this commit, that size reduction would only occur in the case
of a re-ordering. This repairs it to occur whether or not a reordering
occurs.