workspace/snopt: Downgrade to SNOPT 7.4-1.1
When a user chooses to use RobotLocomotion/snopt, now we will compile
the SNOPT 7.4-1.1 2015-02-26 release.  The f2c-generated sources in that
release snapshot brand themselves as 7.4-1.2.

Prior to this commit, we would compile the SNOPT 7.6.1 2018-01-20
release.  The f2c-generated sources in that release brand themselves as
7.4-1.2 as well (no change).

Therefore, this commit should have no effect when using the f2c bindings
for SNOPT (the default).

However, when using the Fortran bindings for SNOPT via Bazel's --config
snopt_fortran option, now we will use the 7.4-era Fortran code, instead
of 7.6.  For now, we prefer that choice so that swapping back and forth
between the two bindings is decoupled from SNOPT's algorithmic changes
that came in with revision 7.6 -- especially because the algorithms in
7.6 have several major regressions compared to 7.4