Remove the "before context allocation" requirement from SceneGraph
SceneGraph previously had documented behavior that all changes must be
made prior to the allocation of a Context. This was enforced with a
runtime test.

This behavior is now deprecated in favor of documenting the difference
between the SceneGraph model and the copy in the context.

It also cleaned up the `@throws` documentation to make it print better in
doxygen (and be compatible with python documentation extraction). The throw
criteria changed, so it made sense to hit the formatting as well.