symbolic: ExtractDoubleOrThrow(NAN) should be NAN (#10773)
During transmogrification of a Context<Expression> to Context<double>,
the user may have placed NAN bombs in state variables or parameters
which should be unused by the computation in question (e.g., the
dynamics may not depend on some output-scaling parameter). In this
case, we should just faithfully convert the Expression NAN to a
double NAN, without throwing an exception.