common: Restrict which RandomGenerator APIs we promise
We should only promote certain items, not offer all mt19337 APIs, so
that we can be clear on the API surface we are promising to maintain,
and clear about the soundness of the APIs our framework relies on.

This also has the advantage that the mangled symbols in our object code
become much thinner for, e.g., primitives::RandomSource, and that the
Value<T>'s TypeHasher can understand how to process a RandomGenerator.