common: Add TypeHasher test for non-type templates
Until C++17 'template <auto...>' arrives, it is somewhat difficult to unpack
non-type template parameter packs.  In the meantime, we can detect this
condition and report a failure to hash.  In practice such templates being used
in Value<T> are rare -- currently there is on only one example.

Also remove 'vector<const double>' unit test case; as it turns out, that
is not a valid template argument for vector in the first place.