common: Use type hashing to make Value be snake_case
The goal here is to remove the notion of accessors being checked or
unchecked with respect to verifying the user-provided T.  All accessors
all checked now, but with a very fast implementation that catches almost
all errors (it uses 64-bit hash mismatches to reject bad Ts) in Release
mode, and all errors in Debug mode.

This commit does not deprecate the old CamelCase methods.  That will
happen later.