Add MultipleShooting::NewSequentialVariable().
This method adds a sequential variable (a variable that has associated decision
variables for each time index) to the optimization problem and returns a
placeholder variable (not actually declared as a decision variable in the
MathematicalProgram). This variable will be substituted for real decision
variables at particular times in methods like AddRunningCost(). The actual
decision variables for a particular time index can be retrieved with
`GetSequentialVariableAtIndex()`.