Maliput: Modifies Intersection to hold a RightOfWayPhaseRing::Id
The Intersection should hold a RightOfWayPhaseRing::Id instead of a
pointer to a RightOfWayPhaseRing because RightOfWayPhaseBook doesn't
provide access to a pointer to a RightOfWayPhaseRing.

If a derivative class of api::Intersection needs access to the
RightOfWayPhaseRing, it can hold a pointer to a RightOfWayPhaseBook,
just like how base::Intersection holds a pointer to a
SimpleRightOfWayPhaseProvider.