Only update engine poses based on cache invalidation
This creates a single cache entry with a meaningless value. Its sole
purpose is to be a cache entry which gets dirtied by its upstream
prerequisites and whose evaluation triggers a calc with *side effects*.
In this case, the side effect is to update the poses of *all* engines.

In the future, this will be broken down further with finer granularity so
that only those components that need updating are done.

(This applies the same mechanism to geometry and geometry/dev.