Querying for anchored world pose now valid
The original bug shown in issue 9145 was for a *different* reason.
Originally, GeometryState segregated anchored and dynamic geometries and
the implementation of GeometryState::get_pose_in_world() didn't consider
both groups. A previous PR merged the geometry types, but failed to update
the newly common registration method to capture the anchored pose.

So, the "same" bug but for different reasons. Now it is properly tested
and the alternative form of the bug has been corrected.