lcm: Add new APIs and discourage use of threads
Add DrakeLcmInterface::HandleSubscriptions to pump the receive queue on
demand (instead of relying on the thread to do it).

Add new Subscriber sugar class to DrakeLcmInterface, where the user
doesn't have to write any lambdas at all.

Add unsubscribe option to DrakeLcmInterface and its sugars.
Add queue_capacity option to DrakeLcmInterface and its sugars.

Add a Doxygen warning to DrakeLcm's thread-management methods as well as
the LcmReceiveThread class.

Condense the implementation of LcmReceiveThread to not duplicate logic
that's already available from the LCM upstream.

Add DrakeLcm::get_lcm_url helper that reports the URL, even if it was
defaulted.