Linearize should allow Abstract inputs.
In reviewing the implementation of additional inputs, I realized that abstract input ports were being handled inconsistently.  If it's ok to have a connected vector input port which gets frozen/cloned, then it should also be ok to have an additional connected abstract input port that behaves the same way.
Also, the implementation in `FixInputPortsFrom` is slightly better than the one that was here (by handling derived types of BasicVector).