Removing pose from GeometryFrame
The GeometryFrame represents a dynamic frame. Evaluating its value will
*always* require connected input ports that will define the frame.
Therefore, initialization is pointless -- the intial value will be stomped
on. This is manifest in the sheer number of sites where the pose is simply
the identity.

Old constructor is marked deprecated in favor of the new constructor which
does not take a pose. All call sites have been updated.