Set "timing level" to zero when calling into snopt.
Before this change, I was noticing a lot of snopt threads (when doing
parallel solves with openmp) sitting in getrusage in backtraces while
working on another issue.

It turns out that the kernel takes a lock when calling `getrusage`
with `RUSAGE_SELF` (see
https://github.com/torvalds/linux/blob/master/kernel/sys.c#L1732 )
which is what's happening in libgfortran, see
https://github.com/gcc-mirror/gcc/blob/master/libgfortran/intrinsics/time_1.h#L149
so we probably shouldn't do that during parallel solves.  For extra
fun, `RUSAGE_SELF` returns the total for all threads, so the
measurement would be garbage anyway.

This change seems to make the issue go away in subsequent backtraces.