Add the abstract RenderEngine
The RenderEngine is the common base-class for all perception-role
renderers. It provides some core functionality for interacting with the
GeometryState (supported by NVI methods).

The interface of the RenderEngine should largely mirror that of the older
RgbdRenderer interface.

This commit excludes the *label* images so that it can be reviewed in
parallel with the RenderLabel PR. Once both have landed, the label image
will be included in the RenderEngine interface.