Uniquify System name in MeshcatPointCloudVisualizer (#11177)
When manually setting the name of MeshcatPointCloudVisualizer, if you
try to add two of these systems to the same DiagramBuilder, it gives an
error saying a system with that name has already been declared.  Since
each system is only designed to visualize one point cloud, it's
reasonable to add multiple systems to a given diagram.