find_resource: Deprecate AddResourceSearchPath
This feature was added to make pydrake paths work, before we had the other
linker-based mechanisms in place to find resources.  (Drake used to have
internal calls to this function.)  Now that we no longer use this in pydrake,
we should remove it for simplicity, and in anticipation of runfiles cleanups.