examples/humanoid_controller: Deprecate the whole package (#11284)
Nobody seems in a hurry to port this away from RigidBodyTree,
or keep up with changes in the systems framework. It is a drag
on development, because it has few tests and running it manually
requires a Gurobi license that few possess; "git remembers" if
we ever want to bring it back.