geometry: Make FrameKinematicsVector less ceremonial
The required call pattern ceremony is difficult to get right when
writing systems that produce frames:

(1) The FlaggedValue (serial numbering) is awkward for output ports to
provide, and unlikely to catch real bugs.  (If we want to catch bugs
related to accidental reuse, then the framework can do that on its own;
in Debug builds, Eval could reset every Calc's output pointer to the
default value prior to calling Calc.)

(2) The IDs can be checked by the SceneGraph based solely on the FrameId.
Requiring that callers copy the SourceId into the output port's value is
useless and painful.