Define a VTK-, OpenGL-based implementation of the RenderEngine interface.
Implementation of the RenderEngine interface using VTK as the details.
This is largely a minimal port form RgbdRenderer to the new SG-compatible
RenderEngine interface.

NOTE: the main `render_engine_vtk.h` is not installed and it is the intent
that people create an instance of the RenderEngineVtk via its factory
method.