Make QueryObject legitimately copyable (#11484)
* Make QueryObject legitimately copyable

This finishes the "live" vs "baked" state of the QueryObject. This
implements the "baked" state -- such that it has an embedded GeometryState
that can support queries but is no longer tied to a changing Context.

NOTE: This doesn't touch dev/query_object.h. The expectation is that it
won't cause problems before it gets deleted.