Move MakeDrakeVisualizerProperties
This utility function was causing other code that used it to end up
dependent on LCM. The solution is to move it out of geometry_visualization.

The movement led to a name change (since it's now presented as a more
general utility.)  So, this changes all the call sites with the new,
generic name.