bugfix - Signed distance queries with a zero threshold lose results
FCL is fundamentally incompatible with a zero-valued threshold. The
broadphase has a *special* function for traversing for distance. This
function assumes that it is only reporting *separating* distance. As such,
it only computes box *separating* distance and doesn't even take action if
the boxes are intersecting.

So, in order to avoid falling into that beartrap, we make sure that we
always pass in a positive tolerance and rely on Drake-side culling to
eliminate any distance results that don't satisfy the threshold.

# Conflicts:
#	geometry/proximity/distance_to_shape.h