[render] Implementation of VTK-based RenderEngine  (#11477)
Introduce the VTK-based RenderEngine into the mainstream code (out of geometry/dev).

This is largely a minimal port form RgbdRenderer to the new SG-compatible
RenderEngine interface.

NOTE: the main `render_engine_vtk.h` is not installed and it is the intent
that people create an instance of the RenderEngineVtk via its factory
method.