Move generic code from MinDistConstraint to MinimumValueConstraint. (#11525)
This adds solvers::MinimumValueConstaint, which ensures that the smallest element
in the vector returned by a user-provided function is above a specified minimum
value. The user-provided function should also take a second argument,
influence_value. Any elements of the return vector that would be greater than
influence_value may safely be ommited.

It also 
- refactors multibody::MinimumDistanceConstraint to use
solvers::MinimumValueConstraint.
- adds  CheckConstraintEvalNonsymbolic() for testing constraints.
- fixes #11538.
- adds a CSS class `pre.unicode-art` that does not add vertical whitespace between lines.