GeometryState adds role removal API. (#11572)
* GeometryState adds role removal API.

This provides the ability to remove one or more roles from a geometry or
from the geometries attached to a frame. It also provides the explicit
functions for removing a geometry from a particular renderer (as opposed
to removing it from all renderers by virtue of removing the perception
role).

Note: some GeometryState documentation refers to implementations of
SceneGraph methods that won't exist until the *next* PR.