Use the event handlers for the receivers
This example overloads the dispatcher functions to handle periodic
discrete update, which is not recommended for general use cases. This PR
switches to use the customized event handlers instead. A forced update
event has been added to guarantee that the updated code is functionally
similar to the original event dispatcher implementation.