Fix error in modifying ProximityProperties
Each modification of properties would duplicate a geometry *in the engine*.
The test has been extended to detect this and the problem has been
alleviated.

The change in geometry_state.cc consists of modifications of todos and
the inclusion of code in an `if () {` block. The actual code hasn't
changed.