MBP "visual" geometry has both illustration and perception roles (#11824)
* MBP "visual" geometry has both illustration and perception roles

In the absence of a robust pipeline for parsing data and assigning distinct
roles, registering a geometry through MBP will cause it to have both
illustration and perception roles.

This hack needs to be removed when it can be done in a more disciplined
way (particularly, when all knowledge of geometry is taken away from MBP).