math: Clarify RigidTransform method deprecation date (#11846)
Because we haven't even marked the C++ methods deprecated yet (and can't
even do so soon, because there is much more porting to do), we should
promise a minimum of four months window.  This is both clearer and less
work than bumping the date to the immediately next month each month.