Updates geometry::Identifier hash method
geometry::Identifier predated the drake hash protocol. It was made to hash
but its hash mechanism apparently precluded it being used in other entities
that *did* use the drake hash protocol.

This changes the hash mechanism to conform and explicitly tests that
SortedPair<GeometryId> is also hashable (as called out in issue 11578).