Correct obj loading in ProximityEngine
If an OBJ referenced a valid mtl file, the effort to parse the file would
fail. This is because we were mis-using the API; we had inadvertantly told
tinyobj to look for the mtl file in the *working* directory instead of the
same directory as the obj file. This makes that change.

Modifies the OBJ used in the test to have a valid mtl file (which caused
the old tests to fail) but pass with the fix.