This creates a simple example for running contact surfaces
It sets up a system that causes changing, intermittent contacts by moving
a sphere, sinusoidally across a box. The sphere is compliant and the box
is rigid.

It includes a system that can translate `ContactSurface` instances into
the a contact path visualization.

It has no dependency on MBP and computes no forces. It is merely a
convenient platform for visualizing contact surfaces. Change the geometry
and observe.