Make DRAKE_EXPECT_THROWS_MESSAGE behave like EXPECT_THROW (#12078)
* Make DRAKE_EXPECT_THROWS_MESSAGE behave like EXPECT_THROW

If DRAKE_EXPECT_THROWS_MESSAGE was used with an incompatible exception type
the test would fail just like it fails if an exception is thrown anywhere
in the code. EXPECT_THROW reports that a different exception type was
thrown and then moves on with the test. Now the drake macro has the same
behavior.

NOTE: It also gets the same behavior as ASSERT_THROW in that it prints
a nice message and then stops the rest of the test -- but it doesn't
crash out of the test.