[geometry] Minimum viable product - ComputeContactSurfaces (#12093)
* Add bare-bones SceneGraph-Hydroelastic support and visualization

This represents the absolute minimum code necessary to query for
ContactSurface from SceneGraph. The full solution is much larger and will
be PR'd into this as we go. The cost of simplification is:

  - Contact surfaces can *only* be computed between soft spheres and rigid
    boxes.
  - All spheres and all boxes are assumed to be compatible.
  - The hydroelastic properties are hard-coded; user configuraiton is a
    follow up feature.
  - The hydroelastic representation is not pre-computed or cached - it is
    generated anew for *every* sphere-box pair.

These failings are all going to be shored up in subsequent PRs.

This also creates a simple example for running contact surfaces

It sets up a system that causes changing, intermittent contacts by moving
a sphere, sinusoidally across a box. The sphere is compliant and the box
is rigid.

It includes a system that can translate `ContactSurface` instances into
the a contact path visualization.

It has no dependency on MBP and computes no forces. It is merely a
convenient platform for visualizing contact surfaces. Change the geometry
and observe.