Add hydroelastic specification to proximity properties
Introduces utilities for writing properties to a ProximityProperties that
SceneGraph depends on: e.g.,  resolution hint, pressure, etc.

Feeds the resolution hint and pressure field to the mesh tesselation and
contact surface evaluation. The effect of resolution hint can be observed
in simple_contact_surface_vis.cc.

This still does not include:
  - Pre-computing hydroelastic representations
  - Utilties for setting more general contact materials (e.g., dissipation,
    coulomb friction, etc.)