Reframe mesh creation with "resolution hint" (#12114)
* Reframe mesh creation with "resolution hint"

We previously had some kind of "edge length" or "target length" in the API
for creating a mesh. After consideration, it was deemed better to make it
more generic as a "resolution hint". Each function takes a resolution hint
as a consistent API and provides a definition of how that hint guides the
tesselation. It is not required to be exactly the same interpretation, but
they should at least be in the same ballpark.