Level-set generated ContactSurface subdivides all faces
Previously, triangles weren't subdivided when the level set computed them.
It is critical to subdivide all faces because small perturbation can change
the "intersection" from a triangle to a quad. Without the centroid fan,
this change will lead to a discontinuity in the quadrature points.

This change adds the fan generation.