Ramp up anti-aliasing in the VTK color image
Previously, MSAA was set to zero on all image types. However, FXAA was
turned on for color images. Apparently, doing so leaves a modicum of
anti-aliasing in the color image (but not much).

This provides a mechanism to specify anti-aliasing (on or off) and links
the two features (MSAA and FXAA) explicitly making a stronger distinction
between turning it on and off (with a supporting test).

By default, color anti aliasing is enabled. The unit test disables it.