Removing grad_H from ContactSurface
We are no longer using the grad_h value to define the normal of the contact
surface. Instead, we're using the triangle normals of the surface itself.
So, we remove the field and change downstream references to use the face
normal instead.

This act revealed some imprecision in the documented and implemented
behaviors of the triangle normals. Where appropriate new documentation has
been added (and tested) to help clarify the expectations of normals.