Adding drake contact material properties to sdf/urdf parsing
This introduces a new XML sub-tree that is largely common between both SDF
and URDF.

SDF version:

  <drake:drake>
    <drake:hydroelastic_resolution_hint>1.3</drake:hydroelastic_resolution_hint>
    <drake:elastic_modulus>1e8</drake:elastic_modulus>
    <drake:dissipation>0.25</drake:dissipation>
    <drake:mu_dynamic>0.7</drake:mu_dynamic>
  </drake:drake>

URDF version:

  <drake:drake>
    <drake:hydroelastic_resolution_hint value="1.3" />
    <drake:elastic_modulus value="1e8" />
    <drake:dissipation value="0.25" />
    <drake:mu_dynamic value="0.7" />
  </drake:drake>

(URDF differs because URDF has a style that uses attributes whereas SDF
eschews them.)

 - Parsing collision geometry now creates a populated ProximityProperties
   and stores the parsed properties there.
 - There is some code that provides legacy compatibility with old
   mechanisms for specifying coefficients of friction so that existing
   files will still work (albeit with warnings).
 - MBP API has been extended to accept proximity properties compatible with
   this parsing logic. The old API needs to be deprecated (in a follow up
   commit).

A note on coefficients of friction. MBP has changed and no longer uses two
coefficients of friction. But we still have CoulombFriction with both
static and dynamic friction. The new "drake:"-prefixed values omit the
static friction for this reason. This PR is *not* changing the
CoulombFriction class as it's beyond this commit's remit. However, it
should be done in the future. About the same time as the legacy friction
parsing is removed. (First it needs to be deprecated.)

We still need to add "soft" deprecation for when we use the legacy
functionality. It can't be compile-time, so it'll have to be warnings with
a time stamp.