geometry: support Mesh in ComputeContactSurfaces().
It is for hydroelastic contact model. Mesh is not supported in other
proximity queries, mostly because we cannot use FCL's mesh
representation, which is a triangle soup without the concept of
enclosing volume.

However, we still need broadphase culling for Mesh geometries colliding
with other geometries. Therefore, we have colliding trees of bounding
boxes of Mesh separated from the trees of other geometries.