Expression(0.0) re-uses the pointer of Expression::Zero()
Given a scalar type T, Eigen assumes that T(0) does not involve memory-allocation. This is not a correct assumption and needs to be fixed, but it takes time to fix the upstream and use the latest version of Eigen. Here, we want to provide a workaround on our side.

See https://github.com/RobotLocomotion/drake/issues/12453.