Pass --no-lock to brew bundle (#12458)
https://github.com/Homebrew/homebrew-bundle/pull/552 introduces
"Brewfile.lock.json" which provides a list of successfully installed packages
and their metainfo.

For now, git detects those new added "Brewfile.lock.json" as untracked
files. Passing `--no-lock` is one solution to this problem. An alternative is to
add the pattern in `.gitignore` but I think passing `--no-lock` is a cleaner
solution.