Expression(0.0) re-uses the pointer of Expression::Zero() (#12459)
Given a scalar type T, Eigen assumes that the constructor for T(0) is
idempotent. Consequently, it should not involve memory-allocation to avoid a
memory leak. This is not a correct assumption and needs to be fixed, but it
takes time to fix the upstream and use the latest version of Eigen. Here, we
want to provide a workaround on our side.

See https://github.com/RobotLocomotion/drake/issues/12453.