Fix address sanitizer problem
`coulomb_friction` accessed a reference into the input properties. Those
properties get *moved* into SceneGraph. However, there is one case where
the moved properties are *discarded* -- when the shape is a mesh. In this
case the captured reference was destroyed leading to referencing freed
memory. This mitigates the immediate symptom.

It also fixes a couple of minor issues found while assessing this issue.