Correct SymbolicVectorSystem "container" population
When evaluating various expressions (output, dynamics, or discrete updates),
SymbolicVectorSystem builds an Environment or Substitution in order to
map inputs to results.

The original implementation would conservatively attempt to populate the
container with every defined variable, even if the expression being
evaluated didn't need it.

This meant if there were *any* input variables, the input port would be
evaluated. In small plant-controller loops, this quickly leads to spurious
algebraic loops as input invokes output and vice versa forever.

This changes the population algorithm to strategically evaluate the input
ports based on a *known* need.