Refactor implicit integration tests
Change the tests in systems/analysis in the following ways in preparation for the new velocity-implicit Euler integrator (#12528):

1. Fix the incorrect velocity in the true solution in StiffDoubleMassSpring, as well as the incorrect damping force calculation in f2 in StiffDoubleMassSpring.
2. Add a velocity solution check to DoubleSpringMassDamperTest.
3. Move common tests from implicit_euler_integrator_test.cc and radau_integrator_test.cc into implicit_integrator_test.h
4. Add two flags into implicit_integrator_test.h to modify or skip tests based on whether an implicit integrator supports Error Control, Autodiff, and Central differencing. By default for existing integrators, all these flags are turned on and the tests are ran in full capability.

The purpose of the fourth change is so that I can separate the new integrator into three different PRs: When the fixed-step integrator is merged, it will run the tests without error control or auto/center diff. When the Error control is merged, it will run the tests with error control. When the full integrator is merged, it'll run the tests with all flags turned on.
By default for other integrators the tests run with full capability. I've tested locally to make sure that my fixed-step version, with error control removed, passes the tests with the flags turned off.