Fix show logic in pyplot visualizer (#12658)
The show arguments passed into the recording methods were doing nothing (they were accidentally setting self.show instead of self._show).  And it was too late, anyhow, to prevent a window being drawn in the jupyter workflow.  This commit does the shenanigans necessary to avoid that extra plot.

Unfortunately, the drawing/not drawing is not something we currently test.