Modifies MBP body -> SG Frame registration (#12690)
* Modifies MBP body -> SG Frame registration

If an MBP is associated with a SG, the goal is to have every body
associated with an SG FrameId, whether it has geometry or not. The current
implementation does this in two ways:

  1. Greedily register a body with SG when a geometry is assigned to it.
  2. At finalize, find any bodies that haven't been registered and register
     them.

This is incompatible with defining collision filter groups. We should be
able to put a body in a collision filter group (even if it doesnt have any
geometry -- it just won't cause anything to be filtered). But that work
flow requires getting that body's FrameId *before* finalization (at least
in how the parsing is currently implemented).

With the same end goal that every body has a frame id, this PR changes the
implementation:

  1. If a SceneGraph is attached, the act of adding a body registers it
     immediately.
  2. When "attaching" the SceneGraph, it looks to see if any bodies were
     added previously and registers them then.

Therefore, we no longer wait for finalization and we're guaranteed that if
there is a SceneGraph attached, every body definitely has a frame.